refactor serde guard json parser
